+++ features/chat/api/remove-collaborator.ts
@@ -1,0 +1,14 @@
+import { trpc } from '@/libs';
+
+/**
+ * Removes a collaborator from a chat
+ */
+export default function useRemoveCollaborator() {
+  const utils = trpc.useContext();
+
+  return trpc.chat.removeCollaborator.useMutation({
+    onSuccess: (_result, input) => {
+      void utils.chat.getCollaborators.invalidate({ chatId: input.chatId });
+    },
+  });
+}