+++ features/chat/routes/get-chat.ts
@@ -2,8 +2,10 @@ import { z } from 'zod';
 
 import { procedure } from '@/server/trpc';
 import { UserRole } from '@/features/shared/types/user';
+import { Forbidden } from '@/features/shared/errors/routeErrors';
 import logger from '@/server/logger';
 import getChat from '@/features/chat/dal/getChat';
+import { getChatAccess } from '@/features/chat/dal/getChatAccess';
 
 const inputSchema = z.object({
   chatId: z.string().uuid(),
@@ -28,10 +30,13 @@ export default procedure
 
     const chat = await getChat(chatId);
 
-    // perform an ownership check with the user's id
-    if (ctx.userRole !== UserRole.Admin && chat.userId !== ctx.userId) {
-      logger.error(`You do not have permission to view this chat: userId: ${ctx.userId}, chatId: ${chat.id}`);
-      throw new Error('You do not have permission to view this chat');
+    // perform an access check with the user's id
+    if (ctx.userRole !== UserRole.Admin) {
+      const access = await getChatAccess(chatId, ctx.userId);
+      if (!access) {
+        logger.error(`You do not have permission to view this chat: userId: ${ctx.userId}, chatId: ${chat.id}`);
+        throw Forbidden('You do not have permission to view this chat');
+      }
     }
 
     return {